% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top2vec.R
\name{top2vecr}
\alias{top2vecr}
\title{Run the top2vecr model}
\usage{
top2vecr(
  x,
  type = "PV-DBOW",
  dim = 300L,
  iter = 40L,
  hs = TRUE,
  window = 15L,
  negative = 0L,
  sample = 1e-05,
  min_count = 50,
  lr = 0.05,
  threads = 4L,
  n_neighbours = 15L,
  n_components = 5L,
  metric = "cosine",
  minPts = 15L
)
}
\arguments{
\item{x}{A `data.frame` object with columns `doc_id` and `text` storing document ids and texts as character vectors}

\item{dim}{Integer value defining the length of embedding vectors}
}
\value{
An object of class ...
}
\description{
Runs an R implementation of the top2vec model as described here: https://arxiv.org/abs/2008.09470.
The R implementation is very similar to the original Python
implementation which can be found here: https://github.com/ddangelov/Top2Vec
}
\examples{
\donttest{
topic_words <- top2vecr(x)
}
}
